requirements.txt
coloredlogs==15.0.1
humanfriendly==10.0
python-dotenv==0.19.2
